Move demo code to a new 'example' folder
